+++ src/steps/create-schema-customization/transform-fields/index.js
@@ -181,7 +181,23 @@ export const transformFields = ({
 
             const aliasedField = source[autoAliasedFieldPropertyName]
 
-            return aliasedField
+            if (typeof aliasedField !== `undefined`) {
+              return aliasedField
+            }
+
+            // the findTypeName helpers was written after this resolver
+            // had been in production for a while.
+            // so we don't know if in all cases it will find the right typename
+            // for this resolver..
+            // So the old way of doing it is above in autoAliasedFieldPropertyName
+            // @todo write comprehesive data resolution integration tests
+            // using many different WPGraphQL extensions
+            // then come back and remove the `return aliasedField` line and
+            // see if this still resolves everything properly
+            const typeName = findTypeName(field.type)
+            const autoAliasedFieldName = `${fieldName}__typename_${typeName}`
+
+            return source[autoAliasedFieldName]
           },
         }
       }