+++ src/steps/source-nodes/fetch-nodes/fetch-nodes-paginated.js
@@ -1,5 +1,6 @@
 import fetchGraphql from "~/utils/fetch-graphql"
 import store from "~/store"
+import { formatLogMessage } from "~/utils/format-log-message"
 
 export const normalizeNode = ({ node, nodeTypeName }) => {
   const normalizedNodeTypeName = node.__typename || nodeTypeName
@@ -22,6 +23,7 @@ const paginatedWpNodeFetch = async ({
   nodeTypeName,
   helpers,
   throwFetchErrors = false,
+  throwGqlErrors = false,
   allContentNodes = [],
   after = null,
   settings = {},
@@ -53,19 +55,22 @@ const paginatedWpNodeFetch = async ({
     variables.first = settings.limit
   }
 
+  const errorContext = `Error occured while fetching nodes of the "${nodeTypeName}" type.`
+
   const response = await fetchGraphql({
     query,
     throwFetchErrors,
+    throwGqlErrors,
     variables: {
       ...variables,
       after,
     },
-    errorContext: `Error occured while fetching nodes of the "${nodeTypeName}" type.`,
+    errorContext,
   })
 
   const { data } = response
 
-  if (!data[contentTypePlural] || !data[contentTypePlural].nodes) {
+  if (!data?.[contentTypePlural]?.nodes) {
     return allContentNodes
   }