+++ src/steps/source-nodes/update-nodes/content-update-interval.js
@@ -2,12 +2,61 @@ import fetchAndApplyNodeUpdates from "./fetch-node-updates"
 import { formatLogMessage } from "~/utils/format-log-message"
 import store from "~/store"
 
-const refetcher = async (msRefetchInterval) => {
-  await fetchAndApplyNodeUpdates({
-    intervalRefetching: true,
-  })
+const refetcher = async (
+  msRefetchInterval,
+  helpers,
+  { reconnectionActivity = null, retryCount = 1 } = {}
+) => {
+  try {
+    await fetchAndApplyNodeUpdates({
+      intervalRefetching: true,
+      throwFetchErrors: true,
+      throwGqlErrors: true,
+    })
 
-  setTimeout(() => refetcher(msRefetchInterval), msRefetchInterval)
+    if (reconnectionActivity) {
+      reconnectionActivity.end()
+      helpers.reporter.success(
+        formatLogMessage(
+          `Content updates re-connected after ${retryCount} ${
+            retryCount === 1 ? `try` : `tries`
+          }`
+        )
+      )
+
+      reconnectionActivity = null
+      retryCount = 1
+    }
+  } catch (e) {
+    if (!reconnectionActivity) {
+      reconnectionActivity = helpers.reporter.activityTimer(
+        formatLogMessage(`Content update error: "${e.message}"`)
+      )
+      reconnectionActivity.start()
+      reconnectionActivity.setStatus(`retrying...`)
+    } else {
+      retryCount++
+      reconnectionActivity.setStatus(`retried ${retryCount} times`)
+    }
+
+    // retry after retry count times 5 seconds
+    const retryTime = retryCount * 5000
+    // if the retry time is greater than or equal to the max (60 seconds)
+    // use the max, otherwise use the retry time
+    const maxWait = 60000
+    const waitFor = retryTime >= maxWait ? maxWait : retryTime
+
+    await new Promise((resolve) => setTimeout(resolve, waitFor))
+  }
+
+  setTimeout(
+    () =>
+      refetcher(msRefetchInterval, helpers, {
+        reconnectionActivity,
+        retryCount,
+      }),
+    msRefetchInterval
+  )
 }
 
 /**
@@ -33,7 +82,7 @@ const startPollingForContentUpdates = (helpers, pluginOptions) => {
     helpers.reporter.info(formatLogMessage`Watching for WordPress changes`)
   }
 
-  refetcher(msRefetchInterval)
+  refetcher(msRefetchInterval, helpers)
 }
 
 export { startPollingForContentUpdates }